RUM-5766 Address comments
