+++ DatadogCore/Tests/DatadogObjc/DDSessionReplayTests.swift
@@ -31,7 +31,7 @@ class DDSessionReplayTests: XCTestCase {
 
     func testConfigurationWithNewApi() {
         // Given
-        let textAndInputPrivacy: DDSessionReplayConfigurationTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
+        let textAndInputPrivacy: DDTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
         let touchPrivacy: DDTouchPrivacyLevel = [.show, .hide].randomElement()!
         let imagePrivacy: DDImagePrivacyLevel = [.maskAll, .maskNonBundledOnly, .maskNone].randomElement()!
         let sampleRate: Float = .mockRandom(min: 0, max: 100)
@@ -56,7 +56,7 @@ class DDSessionReplayTests: XCTestCase {
         // Given
         let sampleRate: Float = .mockRandom(min: 0, max: 100)
         let privacy: DDSessionReplayConfigurationPrivacyLevel = [.allow, .mask, .maskUserInput].randomElement()!
-        let textAndInputPrivacy: DDSessionReplayConfigurationTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
+        let textAndInputPrivacy: DDTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
         let imagePrivacy: DDImagePrivacyLevel = [.maskAll, .maskNonBundledOnly, .maskNone].randomElement()!
         let touchPrivacy: DDTouchPrivacyLevel = [.show, .hide].randomElement()!
         let url: URL = .mockRandom()
@@ -82,7 +82,7 @@ class DDSessionReplayTests: XCTestCase {
     func testConfigurationOverridesWithNewApi() {
         // Given
         let sampleRate: Float = .mockRandom(min: 0, max: 100)
-        let textAndInputPrivacy: DDSessionReplayConfigurationTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
+        let textAndInputPrivacy: DDTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
         let imagePrivacy: DDImagePrivacyLevel = [.maskAll, .maskNonBundledOnly, .maskNone].randomElement()!
         let touchPrivacy: DDTouchPrivacyLevel = [.show, .hide].randomElement()!
         let url: URL = .mockRandom()
@@ -119,13 +119,13 @@ class DDSessionReplayTests: XCTestCase {
     }
 
     func testTextAndInputPrivacyLevelsInterop() {
-        XCTAssertEqual(DDSessionReplayConfigurationTextAndInputPrivacyLevel.maskAll._swift, .maskAll)
-        XCTAssertEqual(DDSessionReplayConfigurationTextAndInputPrivacyLevel.maskAllInputs._swift, .maskAllInputs)
-        XCTAssertEqual(DDSessionReplayConfigurationTextAndInputPrivacyLevel.maskSensitiveInputs._swift, .maskSensitiveInputs)
+        XCTAssertEqual(DDTextAndInputPrivacyLevel.maskAll._swift, .maskAll)
+        XCTAssertEqual(DDTextAndInputPrivacyLevel.maskAllInputs._swift, .maskAllInputs)
+        XCTAssertEqual(DDTextAndInputPrivacyLevel.maskSensitiveInputs._swift, .maskSensitiveInputs)
 
-        XCTAssertEqual(DDSessionReplayConfigurationTextAndInputPrivacyLevel(.maskAll), .maskAll)
-        XCTAssertEqual(DDSessionReplayConfigurationTextAndInputPrivacyLevel(.maskAllInputs), .maskAllInputs)
-        XCTAssertEqual(DDSessionReplayConfigurationTextAndInputPrivacyLevel(.maskSensitiveInputs), .maskSensitiveInputs)
+        XCTAssertEqual(DDTextAndInputPrivacyLevel(.maskAll), .maskAll)
+        XCTAssertEqual(DDTextAndInputPrivacyLevel(.maskAllInputs), .maskAllInputs)
+        XCTAssertEqual(DDTextAndInputPrivacyLevel(.maskSensitiveInputs), .maskSensitiveInputs)
     }
 
     func testImagePrivacyLevelsInterop() {
@@ -171,7 +171,7 @@ class DDSessionReplayTests: XCTestCase {
         // Given
         let core = FeatureRegistrationCoreMock()
         CoreRegistry.register(default: core)
-        let textAndInputPrivacy: DDSessionReplayConfigurationTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
+        let textAndInputPrivacy: DDTextAndInputPrivacyLevel = [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
         let imagePrivacy: DDImagePrivacyLevel = [.maskAll, .maskNonBundledOnly, .maskNone].randomElement()!
         let touchPrivacy: DDTouchPrivacyLevel = [.show, .hide].randomElement()!
         defer { CoreRegistry.unregisterDefault() }
