--- DatadogSessionReplay/Tests/Mocks/TextAndInputPrivacyLevelMocks.swift
@@ -1,20 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2019-Present Datadog, Inc.
- */
-
-#if os(iOS)
-import DatadogSessionReplay
-import TestUtilities
-
-extension TextAndInputPrivacyLevel: AnyMockable, RandomMockable {
-    public static func mockAny() -> Self {
-        .maskSensitiveInputs
-    }
-
-    public static func mockRandom() -> Self {
-        [.maskAll, .maskAllInputs, .maskSensitiveInputs].randomElement()!
-    }
-}
-#endif
